### 2024-01-19 #6246 @editorialbot
[PRE REVIEW]: The fqar package: R tools for analyzing floristic quality assessment data
**Submitting author:** @equitable-equations (<a href="http://orcid.org/0000-0003-4434-0755">Andrew Gard</a>)
**Repository:** https://github.com/equitable-equations/fqar
**Branch with paper.md** (empty if default branch): joss
**Version:** v0.5.1
