### 2024-01-19 #6246 @kthyng
Hi @equitable-equations and thanks for your submission. Can you link to your previous JOSS paper on this software so I can compare?
