### 2024-01-26 #6246 @equitable-equations
Apologies, I didn't see your request amid the various editorialbot comments. The previous iteration of this paper was desk rejected for insufficient scholarly content (small size). At the time, none of the package's analytical functionality had been implemented and the codebase was only ~800 lines.

As for the new paper itself, the introduction/background on FQA is very similar but the later sections (especially Workflow) have been updated to appropriately emphasize the package's statistical co-occurrence functions. The JOSS link is no longer active; I'll send my own old draft later in the day.
