+++ paper/fqar_refs.bib
@@ -1,0 +1,186 @@
+%% This BibTeX bibliography file was created using BibDesk.
+%% https://bibdesk.sourceforge.io/
+
+%% Created for Andrew Gard at 2023-08-10 14:35:11 -0500 
+
+
+%% Saved with string encoding Unicode (UTF-8) 
+
+
+
+@inproceedings{wilhelm2017flora,
+	author = {Wilhelm, Gerould and Rericha, Laura},
+	date-added = {2023-08-10 14:34:57 -0500},
+	date-modified = {2023-08-10 14:34:57 -0500},
+	organization = {Indiana Academy of Science Indianapolis},
+	title = {Flora of the Chicago region: A floristic and ecological synthesis},
+	year = {2017}}
+
+@article{oldham1995floristic,
+	author = {Oldham, MJ and Bakowsky, WD and Sutherland, DA},
+	date-added = {2022-08-30 10:59:12 -0500},
+	date-modified = {2022-08-30 10:59:12 -0500},
+	journal = {Ontario. Natural Heritage Information Center Report, Ministry of Natural Resources, Peterborough, Ontario, Canada},
+	title = {Floristic Quality Assessment System},
+	year = {1995},
+	bdsk-url-1 = {http://dx.doi.org/10.13140/RG.2.2.35685.91360}}
+
+@article{freeman2012coefficients,
+	author = {Freeman, Craig C},
+	date-added = {2022-08-30 10:23:15 -0500},
+	date-modified = {2022-08-30 10:23:15 -0500},
+	journal = {Kansas Biological Survey \& RL McGregor Herbarium http://ksnhi. ku. edu/resources/plants},
+	title = {Coefficients of conservatism for Kansas vascular plants (2012) and selected life history attributes},
+	year = {2012}}
+
+@article{bowles2006testing,
+	author = {Bowles, Marlin and Jones, Michael},
+	date-added = {2022-08-30 10:07:03 -0500},
+	date-modified = {2022-08-31 17:49:04 -0500},
+	journal = {Natural Areas Journal},
+	number = {1},
+	pages = {17--30},
+	publisher = {BioOne},
+	title = {Testing the efficacy of species richness and floristic quality assessment of quality, temporal change, and fire effects in tallgrass prairie natural areas},
+	url = {https://doi.org/10.3375/0885-8608(2006)26[17:TTEOSR]2.0.CO;2},
+	volume = {26},
+	year = {2006},
+	bdsk-url-1 = {https://doi.org/10.3375/0885-8608(2006)26%5B17:TTEOSR%5D2.0.CO;2}}
+
+@article{bauer,
+	author = {Bauer, J., Koziol, L., & Bever, J.},
+	date-added = {2022-08-30 09:42:59 -0500},
+	date-modified = {2022-08-30 09:42:59 -0500},
+	journal = {AoB PLANTS},
+	title = {Ecology of Floristic Quality Assessment: testing for correlations between coefficients of conservatism, species traits and mycorrhizal responsiveness},
+	url = {https://doi.org/10.1093/aobpla/plx073},
+	volume = 10,
+	year = 2018,
+	bdsk-url-1 = {https://doi.org/10.1093/aobpla/plx073}}
+
+@article{zinnen,
+	author = {Zinnen, J., Spyreas, Zaya, D., Matthews, J.},
+	date-added = {2022-08-30 09:42:52 -0500},
+	date-modified = {2022-08-30 10:16:57 -0500},
+	journal = {Ecological Indicators},
+	title = {Niche ecology in Floristic Quality Assessment: Are species with higher conservatism more specialized?},
[…]
