### 2024-01-26 #6246 @equitable-equations
Hi,
The previous version is attached and can be found at https://github.com/equitable-equations/fqar/tree/joss_old.
Thanks for your work.

[paper.pdf](https://github.com/openjournals/joss-reviews/files/14069912/paper.pdf)
