### 2024-01-29 #6246 @kthyng
@equitable-equations Ah I meant the JOSS pre-review issue from when you submitted before. Can you dig that up and attach?
