### 2024-02-17 #6366 @mikemahoney218
👋🏼 @equitable-equations, @ifoxfoot, @mhesselbarth : this is the review thread for the paper. Just about all of our communications will happen here from now on 😄

```@editorialbot generate my checklist```

as the top of a new comment in this thread. For best results, don't include anything else in the comment!

This will create a checklist that walks through the JOSS submission requirements. As you go over the submission, please check any items that you feel have been satisfied. The first comment in this thread also contains links to the JOSS reviewer guidelines.

We aim for reviews to be completed within about 4 weeks. Please let me know if you require some more time.

Please feel free to ping me (@mikemahoney218) if you have any questions/concerns. Thanks again so much for agreeing to review!
