### 2024-02-20 equitable-equations/fqar#7 @ifoxfoot
Unable to replicate results in openjournals/joss-reviews#6366 article proof
openjournals/joss-reviews/issues/6366

Hello. I am able to install the `fqar` package both via github and CRAN. I believe I am using the package downloaded from CRAN (they both have the same version number so I can't tell).

I am able to run the code `chicago_invs <- assessment_list_inventory(chicago_fqas)` which returns a large list with 309 elements.

However, when I run the next line of code `chicago_cooccurrence <- assessment_cooccurrences(chicago_invs)` I get a dataframe with 0 observations and 9 variables.

All of the rest of the code in the article proof gives errors.

My internet is not the greatest and it did cut out twice during the download of the Chicago fqas. Also my R version is a little behind. See session info below. Let me know if that could be the issue? Thanks!

````
> sessionInfo()
R version 4.3.0 (2023-04-21 ucrt)
Platform: x86_64-w64-mingw32/x64 (64-bit)
Running under: Windows 10 x64 (build 19044)

Matrix products: default

locale:
[1] LC_COLLATE=English_United States.utf8  LC_CTYPE=English_United States.utf8
[3] LC_MONETARY=English_United States.utf8 LC_NUMERIC=C
[5] LC_TIME=English_United States.utf8

time zone: America/Los_Angeles
tzcode source: internal

attached base packages:

other attached packages:
[1] fqar_0.5.1

 [1] vctrs_0.6.2      httr_1.4.7       cli_3.6.1        rlang_1.1.1      purrr_1.0.1
 [6] generics_0.1.3   jsonlite_1.8.8   glue_1.6.2       colorspace_2.1-0 scales_1.3.0
[11] fansi_1.0.4      grid_4.3.0       munsell_0.5.0    tibble_3.2.1     fastmap_1.1.1
[16] lifecycle_1.0.4  memoise_2.0.1    compiler_4.3.0   dplyr_1.1.2      pkgconfig_2.0.3
[21] tidyr_1.3.1      rstudioapi_0.14  R6_2.5.1         tidyselect_1.2.0 utf8_1.2.3
[26] pillar_1.9.0     magrittr_2.0.3   tools_4.3.0      gtable_0.3.4     ggplot2_3.4.2
[31] cachem_1.0.8
````
