### 2024-02-20 equitable-equations/fqar#7 @equitable-equations
@ifoxfoot thanks for reaching out, as well as for your helpful notes on the paper itself, which I'll reply to as soon as possible. I was just able to run those commands successfully on my Macbook, with `chicago_cooccurrence <- assessment_cooccurrences(chicago_invs)` giving a data frame with 4,463,025 rows.

If the later commands (species_profile and species_profile_plot) are giving errors as well, then something has likely gone wrong at the downloading step. We've worked hard to make our functions robust against interruptions, but it's totally possible something has slipped. I'll do some more digging.

Again, thank you for your work!
