### 2024-02-20 equitable-equations/fqar#7 @ifoxfoot
Hi @equitable-equations I'm also testing it on my work computer which has some restrictions--I don't have admin status. I can try it on my personal computer with a more updated R version soon. That could also be the issue. Thank you for your prompt reply!
