### 2024-02-21 equitable-equations/fqar#7 @equitable-equations
Another thing is that we intentionally chose a large and popular FQA database when writing this paper, which has the side effect of making the code slower and the outputs very large. In other words, we weren't really thinking about readers copy/pasting our code. We could consider using a smaller example which would run more reliably on older systems.
