### 2024-02-21 equitable-equations/fqar#7 @ifoxfoot
Update: I mistyped in my orignal issue. I get a large list from the line `chicago_fqas <- download_assessment_list`. I *was* getting an empty data frame from the line `chicago_invs <- assessment_list_inventory(chicago_fqas)`. Sorry about that!

I ran the code with no internet interruptions and it worked on the same system so that is good.

Then I ran it again and interrupted the connection by disconnecting and connecting to the internet. The console looks like this

````
> chicago_fqas <- download_assessment_list(database_id = 80)
  |=====                                                                                                                                                                                                                                                                                                                |   2%Unable to connect. Please check internet connection.
  |=====================================================================================================================================================================================================================================================================================================================| 100%
````
Again I was able to download a large list with 309 elements. It looks like a successful download but it must not be.

then when I run the next line of code I get

````
chicago_invs <- assessment_list_inventory(chicago_fqas)
assessment_list must be a list of dataframes obtained from universalFQA.org. Type ?download_assessment_list for help.
````
Also `chicago_invs` is a df with 0 observations and 9 variables.

So I think the issue is that when the download function has a lapse in internet connection it cannot recover. It's not a huge issue so I will continue with the review. Thanks!
