### 2024-02-21 equitable-equations/fqar#7 @equitable-equations
Ok, great. I'm adding this to my to-do list for the next update. Thanks!
