### 2024-03-07 #6366 @equitable-equations
@mikemahoney218 I already have paper revision requests and suggestions from @ifoxfoot which I'm in the process of implementing or otherwise addressing. My thanks to all concerned!
