### 2024-03-09 equitable-equations/fqar#6 @equitable-equations
@ifoxfoot thank you for these thoughtful comments! I've taken them into account and posted a revision [here](https://github.com/equitable-equations/fqar/tree/joss_revision/paper). Some specific responses to your points:

### Lines 6-7: I would use a different citation.

Response: The recommended citation has been added.

### Lines 7-8: I would reword this sentence to be more factually correct.

Response: A parenthetical has been added to explicitly mention that non-native species are assigned a C-value of 0.

### Lines 12-15: Also, why not cite Swink and Wilhelm 1994 for the floristic quality index metric.

Response: The Wilhelm citation has been added earlier, as suggested. At this point in the paper, we feel that the Zinnen reference is appropriate since it explicitly describes how FQA is actually used in the field today.

### Lines 30-32: Clarify what an assessment is and clarify the purpose/utility of the universal calculator.

Response: This paragraph has been recast to better reflect the utility of the universalFQA calculator.

### Lines 40-41: I'm not sure you can determine if a species is misclassified based on this type of analysis.

Response: In collaboration with local botanists, the authors are in the process of considering exactly this question (and others like it) for a forthcoming paper. Our general premise is that as long as two species are found simultaneously by a practitioner who chooses to include them both in a single site assessment, they may reasonably be taken to be co-occurring. We recognize the potential limitations of such studies but are also finding them to be both useful and insightful.

A spacial analysis would be incredibly valuable and is an important direction for future study. Please let me know if you'd like to collaborate on this!
