### 2024-03-11 #6366 @ifoxfoot
hi @mikemahoney218 I plan on going through the documentation and reading @equitable-equations's updates to the manuscript today! I should be finished with the checklist in a couple of days max!
