### 2024-03-11 #6366 @mikemahoney218
Also, looking at the paper preview, I _think_ that any changes @equitable-equations makes won't be reflected in the preview until they get merged into the `joss` branch. As a result, I think the above preview reflects the original submission, and not the `joss_revisions` branch. Just want to make sure there's no confusion why changes aren't being reflected!
