### 2024-03-11 #6366 @equitable-equations
@mikemahoney218 that's right. I wasn't sure about standard practice on this so I made a new temporary branch while others are still working. The changes are small, but if anyone wants to see the updated version directly you can do so [here](https://github.com/equitable-equations/fqar/tree/joss_revision).
