### 2024-03-11 #6366 @ifoxfoot
@mikemahoney218 and @equitable-equations okay thanks! That makes sense.

I do have one question for @mikemahoney218. Both the documentation and the software paper sections have checklist items for a statement of need. Is it necessary to have a statement of need in two places? Or is one statement of need in the paper sufficient to check both items off? Thanks!
