### 2024-03-11 #6366 @mikemahoney218
We're generally looking for a statement of need in both the paper and the documentation. The main idea is that most users (and other people encountering the project) are probably going to find the documentation before the paper, and it's useful to give them a sense of what your project does and why. That can make it a lot easier to understand the intended use cases for a project and onboard new users.

That said, this section doesn't need to be as formal as it would be in the paper -- to use an example from one of my own projects, I generally include a "Why do this?" in the Readme:
https://github.com/ropensci/unifir?tab=readme-ov-file#why-do-this
