### 2024-03-11 equitable-equations/fqar#8 @ifoxfoot
Comments on R package documentation
openjournals/joss-reviews#6366

# README
- It looks like there are two periods after the link to the universal calculator in the first paragraph.
- You present four types of functions: indexing, downloading, tidying, and analytic. For the last three types there is some text explaining what the functions do interspersed with code blocks, but for the indexing functions there is only code. Could you add some text explaining what the functions do? Since all of your comments in the code block start with download, can you explain how the functions are different from downloading functions in the next section?
- One item in the checklist is "Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support." You could add community guidelines to the readme. There is already a little info about seeking support at the end.

# Vignette
- There are only two authors listed on the vignette but there are three on the paper/R package. Just wondering if this is on purpose.

# R package documentation
- In the description file lines 11-13, you could change the description to specify that you are downloading floristic quality assessment data *from the universal FQA calculator.*
- Documenting what data goes in each column of your function outputs would be helpful for all of your functions. Some column names are self explanatory but some are not. For example, what is `discrepancy_c` in the output of `assessment_cooccurrences_summary`?
- In the function documentation of `assessment_glance` the description of the data returned is different from all of your other functions.  In your other function documentation you write
```
\itemize{
    \item title (character)
...
```
but in the `assessment_glance` documentation you have
```
\itemize{
    title (character)
...
```
so when it is rendered it looks slightly different. Just FYI

# General Documentation Comments
- One thing you could mention is who hosts the Universal FQA Calculator database.
- For arguments like `database_id`, you mention that the ID is specified by the universal calculator. Is there any risk that these IDs would change over time? When a new database is added does it get the next largest ID number? This would also apply to `assessment_id` and `transect_id`
