### 2024-03-13 equitable-equations/fqar#9 @ifoxfoot
Importing utils causes data viewer issue
https://github.com/openjournals/joss-reviews/issues/6366

I'm getting this issue https://github.com/rstudio/rstudio/issues/14009 when I load fqar using `devtools::load_all(".")`. It does not happen when I load the package with `library(fqar)`. It looks like it is caused by importing utils. It seems like you only use utils for the progress bar in some downloading functions, maybe you could import those specific functions instead?

In case this is an issue to specific to my computing environment:
R version 4.2.3 (2023-03-15)
Platform: x86_64-apple-darwin17.0 (64-bit)
Running under: macOS Monterey 12.7.2
