### 2024-03-15 equitable-equations/fqar#10 @mhesselbarth
Adding how to contribute instructions
openjournals/joss-reviews/issues/6366

Please add some "how to contribute" guidelines. For example, you could use `usethis::use_tidy_contributing()`.
