### 2024-03-15 equitable-equations/fqar#11 @mhesselbarth
Adding information about the ID
https://github.com/openjournals/joss-reviews/issues/6366

It think it would be great to add some more information (very brief should be enough though) what the `database_id` is to the README so new users know how to find it when downloading data on their own.

(Similar as in the help file of `download_assessment_list`)
