### 2024-03-15 equitable-equations/fqar#12 @mhesselbarth
(Optional) Example flags
https://github.com/openjournals/joss-reviews/issues/6366

This one is optional, but make sure you actually want to use `\donttest{}` and not `\dontrun{}`. As far as I understand it, the first only doesn't run tests on CRAN, while the later really doesn't run the examples. In my opinion, `\dontrun{}` is much more common, which is why I was a bit confused by the `\donttest{}`
