### 2024-03-15 equitable-equations/fqar#13 @mhesselbarth
Optional link to Vignette
https://github.com/openjournals/joss-reviews/issues/6366

This one is optional, but it would be nice to host the very nice vignette as a rendered homepage (e.g., using `pkgdown`).
