### 2024-03-15 #6366 @mhesselbarth
I finally started with the review and opened a few issues, mostly minor stuff - @equitable-equations et al. very nice packages!

@mikemahoney218 I was trying to rebuild the PDF from the `joss-revision` branch to work on top of @ifoxfoot comments and read the most recent version. Could you please assist given that there seems to be some error building
