### 2024-03-15 equitable-equations/fqar#14 @mhesselbarth
Manuscript
https://github.com/openjournals/joss-reviews/issues/6366

Line 8: Given that non-native species are inlcuded by scored 0, I would reword the sentence and remove the first "native"

Line 41: How would a misclassified species be identified? Only by using certain outliers assuming they are wrong? However, they could also be "true" outliers, correct?

Misc: How does the package relate to other, existing packages, such as https://cran.r-project.org/web/packages/fqacalc/index.html?
