### 2024-03-15 equitable-equations/fqar#14 @equitable-equations
@mhesselbarth thanks for your work! I'll revise to reflect these and other points over the weekend.
