### 2024-03-17 equitable-equations/fqar#13 @equitable-equations
I totally agree. This will happen over the summer, if the creek don't rise.
