### 2024-03-17 equitable-equations/fqar#11 @equitable-equations
Done. The comment preceding the `download_assessment` example now reads:
> download a single assessment using the `assessment_id` assigned by
> [universalfqa.org](https://universalfqa.org/). These identifiers
> can be found using `index_fqa_assessments`.
Good suggestion!
