### 2024-03-17 equitable-equations/fqar#12 @equitable-equations
Yeah, this is the result of specific instructions from the initial human CRAN reviewer who didn't like the use of `dontrun`, possibly because I'd applied it to *all* downloading functions at that time. Honestly I just did what CRAN said.

The package has since been updated to be more stable under internet interruptions, so this can be probably be changed without causing them too much frustration. I agree it would be a positive update and I'll explore it.
