### 2024-03-18 equitable-equations/fqar#12 @mhesselbarth
Again, I don't really know which one is better in which situation. I just hardly never se the `donttest` flag. However, if this was instructed by the CRAN team it's probably the best to just stick with it and ignore my comment :)
