### 2024-03-18 equitable-equations/fqar#15 @ifoxfoot
[optional] suggested improvements to testing
https://github.com/openjournals/joss-reviews/issues/6366

Overall I think your tests are pretty good and have good coverage! I have a couple of suggestions for improvement.

First, there is a typo in name of test file and in title of `test_that(...` for the `assessment_cooccurrences_summary` function. It should be `assessment_cooccurrences_summary` with two Os.

Second, It looks like some of your tests don't actually test that the entire product of the function is correct, instead they just test that the column names and types are correct. Considering your functions often return an empty df when they fail, these types of tests can pass when the function fails.

For example, in the test for `assessment_cooccurences_summary` in line 18 you can change

```
test_cooccur <- assessment_cooccurrences_summary(good_list)
```
to

```
test_cooccur <- assessment_cooccurrences_summary("text")
```
and it still passes.

It looks like in other tests you check for number of rows, which is good. I think it would improve the strength of your tests if you at least check that some of the values are correct and the function is producing the correct number of rows for every output.
