### 2024-03-24 #6366 @equitable-equations
Thanks everyone for your patience. The API on which this package relies is currently failing in a way not anticipated by my tests (certain calls cause it to freeze with no error or even timeout). I've notified the developer and am working to update my test to account for this behavior. I'll respond to the rest of your suggestions as soon as this is resolved.
