### 2024-03-24 equitable-equations/fqar#15 @equitable-equations
I totally agree with this comment. I'll be working with a collaborator to improve the testing suite in June. This should lead to usability and performance enhancements as a side effect - fingers crossed.

I've fixed that typo in the development version and the change will be reflected in the next CRAN update. Thanks for being so thorough!
