+++ tests/testthat/test-database_inventory.R
@@ -13,10 +13,15 @@ test_that("database_inventory works", {
 
   skip_if_offline()
 
-  test_raw <- download_assessment(25002)
-  test_auto <- assessment_inventory(test_raw)
-  expect_equal(ncol(test_auto), 9)
-  expect_equal(names(test_auto)[5], "c")
-  expect_equal(typeof(test_auto$c), "double")
+  test_raw <- suppressMessages(download_assessment(25002))
 
+  if (nrow(test_raw != 0)) {
+    # when server responds
+    test_auto <- database_inventory(test_raw)
+    expect_equal(ncol(test_auto), 9)
+    expect_equal(names(test_auto)[5], "c")
+    expect_equal(typeof(test_auto$c), "double")
+  } else {
+    expect_message(database_inventory(test_raw))
+  }
 })
