+++ tests/testthat/test-species_c.R
@@ -16,17 +16,21 @@ test_that("species_c works", {
 
   skip_if_offline()
 
-  db <- download_database(1)
-  db_inv <- database_inventory(db)
+  db <- suppressMessages(download_database(1))
+  db_inv <- suppressMessages(database_inventory(db))
 
   expect_error(species_c(species, 149, db_inv),
                "database_id or database_inventory cannot both be specified.")
-  expect_message(species_c("fake_species", database_inventory = db_inv),
-               "Species not found in specified database.")
   expect_true(is.na(suppressMessages(species_c("fake_species",
                               database_inventory = db_inv))))
 
-  expect_equal(species_c(species, 149), 4)
-  expect_equal(species_c(species2, database_inventory = db_inv), 10)
+  if (!is.na(suppressMessages(species_c(species, 149)))) {
+    expect_equal(species_c(species, 149), 4) # if server responds
+    expect_equal(species_c(species2, database_inventory = db_inv), 10)
+    expect_message(species_c("fake_species", database_inventory = db_inv),
+                   "Species not found in specified database.")
+  } else {
+    expect_message(species_c(species, 149))
+  }
 
 })
