+++ tests/testthat/test-is_transect_list.R
@@ -1,10 +1,7 @@
 test_that("is_transect_list works",  {
 
-  skip_if_offline()
-
   test_raw <- suppressMessages(download_transect(4492))
   test_transect_list <- list(test_raw, test_raw)
-  test_inv_list <- suppressMessages(transect_list_inventory(test_transect_list))
   bad_list <- list(faithful, faithful)
 
   expect_false(is_transect_list(random_unassigned_variable_5000))
