+++ tests/testthat/test-is_transect.R
@@ -7,8 +7,6 @@ test_that("is_transect works", {
   expect_false(is_transect("hi"))
   expect_false(is_transect(data.frame()))
 
-  skip_if_offline()
-
   test_raw <- suppressMessages(download_transect(4492))
 
   expect_true(is_transect(test_raw))
