### 2024-03-31 fdc5712
improved stability when server is down
