+++ tests/testthat/test-download_database_internal.R
@@ -10,13 +10,11 @@ test_that("download_database_internal works", {
   test_db <- suppressMessages(download_database_internal(1))
 
   if (nrow(test_db) != 0) {
+    # when server responds
     expect_equal(ncol(test_db), 9)
     expect_equal(test_db$V1[1], "Chicago Region")
     expect_equal(class(test_db), c("tbl_df",
                                    "tbl",
                                    "data.frame"))
-  } else {
-    # for when server doesn't respond
-    expect_message(download_database_internal(1))
   }
 })
