+++ tests/testthat/test-assessment_glance.R
@@ -15,13 +15,14 @@ test_that("assessment_glance works", {
   expect_equal(nrow(suppressMessages(assessment_glance(NULL))), 0)
   expect_message(assessment_glance(NULL), "data_set is NULL. Empty data frame returned.")
 
-  skip_if_offline()
-
   test_auto <- suppressMessages(download_assessment(25002))
   test <- suppressMessages(assessment_glance(test_auto))
 
   expect_equal(ncol(test), 52)
-  expect_equal(names(test)[8], "fqa_db_region")
-  expect_equal(names(test)[42], "grass")
-  expect_equal(typeof(test$total_mean_c), "double")
+
+  if (nrow(test_auto) != 0){
+    expect_equal(names(test)[8], "fqa_db_region")
+    expect_equal(names(test)[42], "grass")
+    expect_equal(typeof(test$total_mean_c), "double")
+  }
 })
