+++ tests/testthat/test-download_assessment.R
@@ -7,13 +7,13 @@ test_that("download_assessment works", {
   expect_equal(nrow(null_output), 0)
   expect_equal(memoise::has_cache(download_assessment_internal)(-40000), FALSE)
 
-  skip_if_offline()
-
   test_assessment <- suppressMessages(download_assessment(25002))
   expect_equal(ncol(test_assessment), 9)
 
   if (nrow(test_assessment) !=0){
   expect_equal(memoise::has_cache(download_assessment_internal)(25002), TRUE)
+  expect_equal(nrow(test_a), 140)
+  expect_equal(test_a$V1[1], "Edison dune and swale")
   }
 })
 
