### 2024-04-08 #6366 @equitable-equations
@mikemahoney218 good timing. I'm pushing an update to CRAN shortly to address these new potential behaviors from the API. I have responses to other edit requests prepared and will post them for reviewers as soon as the new version is up.
