+++ tests/testthat/test-download_assessment.R
@@ -12,8 +12,8 @@ test_that("download_assessment works", {
 
   if (nrow(test_assessment) !=0){
   expect_equal(memoise::has_cache(download_assessment_internal)(25002), TRUE)
-  expect_equal(nrow(test_a), 140)
-  expect_equal(test_a$V1[1], "Edison dune and swale")
+  expect_equal(nrow(test_assessment), 140)
+  expect_equal(test_assessment$V1[1], "Edison dune and swale")
   }
 })
 
