### 2024-04-09 6cbb992
add donttest for slow exampl
