### 2024-04-09 669e714
update news for 0.5.2
