+++ man/download_assessment.Rd
@@ -32,7 +32,9 @@ databases <- index_fqa_databases() # Database 1 is the original 1994 Chicago edi
 chicago_assessments <- index_fqa_assessments(1) # Edison dune and swale has id number 25002.
 edison <- download_assessment(25002)
 
+donttest{
 edison_tidy <- assessment_glance(edison)
 edison_species <- assessment_inventory(edison)
+}
 
 }
