### 2024-04-09 bdcc8ef
add donttest for slow example
