### 2024-04-09 ea19cc8
add donttest for slow example
