+++ R/download_assessment.R
@@ -20,12 +20,13 @@
 #' @importFrom memoise drop_cache
 #'
 #' @examples
+#'
+#' \donttest{
 #' databases <- index_fqa_databases() # Database 1 is the original 1994 Chicago edition.
 #'
 #' chicago_assessments <- index_fqa_assessments(1) # Edison dune and swale has id number 25002.
 #' edison <- download_assessment(25002)
 #'
-#' \donttest{
 #' edison_tidy <- assessment_glance(edison)
 #' edison_species <- assessment_inventory(edison)
 #' }
