+++ cran-comments.md
@@ -1,22 +1,15 @@
 ## R CMD check results
 
-0 errors | 0 warnings | 1 note
+0 errors | 0 warnings | 0 notes*
 
-* This packages was archived due to examples that threw an error when run without internet connectivity. I previously communicated with Uwe about this issue. Regardless, the relevant functions have now been re-structured so that all examples and tests run without internet. I apologize for any extra work this might have caused you!
-
-* Additionally, rhub::check_for_cran sometimes produces the following notes, which I understand to be noise:
+* rhub::check_for_cran sometimes produces the following notes, which I understand to be noise:
 
 1. Found the following files/directories: 'lastMiKTeXException'
 
-2. checking for non-standard things in the check directory ''NULL''
+2. Checking for non-standard things in the check directory ''NULL''
 
-## CRAN checks from https://cran.r-project.org/web/packages/fqar/index.html
+* A website referred to in the documentation has been down occasionally in recent weeks. This may possibly produce another note. I've chose to leave the link in the help files for added transparency. 
 
-I have corrected the issues listed on this page and taken step to prevent their recurrence. 
-
-* Thank you for your work!!
-
-## Downstreatm dependencies
 
 There are currently no downstream dependencies for this package.
 
