### 2024-04-09 17629d3
improved handling of down API
