+++ man/download_assessment.Rd
@@ -27,12 +27,13 @@ numbers for assessments in various databases can be found using the
 \code{\link[=index_fqa_assessments]{index_fqa_assessments()}} function.
 }
 \examples{
+
+\donttest{
 databases <- index_fqa_databases() # Database 1 is the original 1994 Chicago edition.
 
 chicago_assessments <- index_fqa_assessments(1) # Edison dune and swale has id number 25002.
 edison <- download_assessment(25002)
 
-\donttest{
 edison_tidy <- assessment_glance(edison)
 edison_species <- assessment_inventory(edison)
 }
