+++ tests/testthat/test-download_transect_list.R
@@ -8,7 +8,7 @@ test_that("download_transect_list works", {
 
   two_transects <- suppressMessages(download_transect_list(1, id == 6570 | id == 6322))
 
-  if (length(two_transects !=0)){
+  if (length(two_transects) !=0){
     expect_equal(class(two_transects), "list")
     expect_equal(length(two_transects), 2)
     expect_equal(class(two_transects[[1]]), c("tbl_df",
