+++ tests/testthat/test-transect_list_glance.R
@@ -12,8 +12,8 @@ test_that("transect_list_glance works", {
 
   test_vec <- c(6875, 6736)
 
-  test_list <- if (length(suppressMessages(download_transect_list(63, id %in% test_vec))) != 0){
-
+  if (length(suppressMessages(download_transect_list(63, id %in% test_vec))) != 0){
+    test_list <- download_transect_list(63, id %in% test_vec)
     test_df <- transect_list_glance(test_list)
 
     expect_equal(ncol(test_df), 54)
