+++ tests/testthat/test-transect_list_glance.R
@@ -24,3 +24,4 @@ test_that("transect_list_glance works", {
     expect_message(download_transect_list(63, id %in% test_vec))
   }
 })
+
