### 2024-04-11 7832b8e
improved testing for API failure
