### 2024-04-11 ff3d0a0
improved testing when API fails
