+++ tests/testthat/test-download_assessment_list.R
@@ -2,8 +2,9 @@ test_that("download_assessment_list works", {
 
   expect_equal(class(suppressMessages(download_assessment_list(-2))), "list")
 
-  two_assessments <- suppressMessages(download_assessment_list(1,
-                                                               id == 8 | id == 12))
+  two_assessments <- suppressMessages(
+    download_assessment_list(1,id == 8 | id == 12)
+    )
   if(length(two_assessments) != 0) {
     # when server responds
     expect_message(download_assessment_list(1, id == "hi"), "No matches found. Empty list returned.")
@@ -12,9 +13,6 @@ test_that("download_assessment_list works", {
     expect_equal(class(two_assessments[[1]]), c("tbl_df",
                                                 "tbl",
                                                 "data.frame"))
-  expect_equal(ncol(two_assessments[[1]]), 9)
-  } else {
-    # if server doesn't respond
-    expect_message(download_assessment_list(1))
+    expect_equal(ncol(two_assessments[[1]]), 9)
   }
 })
