+++ tests/testthat/test-index_fqa_databases.R
@@ -2,9 +2,7 @@ test_that("index_fqa_databases works", {
 
   expect_error(index_fqa_databases("hi"))
 
-  skip_if_offline()
-
-  databases <- index_fqa_databases()
+  databases <- suppressMessages(index_fqa_databases())
   expect_equal(ncol(databases), 4)
   expect_equal(names(databases), c("database_id", "region", "year", "description"))
   expect_equal(class(databases[[1]]), "numeric")
