+++ tests/testthat/test-species_nativity.R
@@ -11,8 +11,6 @@ test_that("species_nativity works", {
   expect_error(species_nativity(species, database_id = "hi"),
                "database_id must be an integer.")
 
-  skip_if_offline()
-
   db <- suppressMessages(download_database(1))
   db_inv <- suppressMessages(database_inventory(db))
 
@@ -21,14 +19,9 @@ test_that("species_nativity works", {
   expect_true(is.na(suppressMessages(species_nativity("fake_species",
                                                database_inventory = db_inv))))
 
-  if (!is.na(suppressMessages(species_nativity(species, 149)))) {
-    expect_equal(species_nativity(species, 149), "native")
-    expect_equal(species_nativity(species3, 149), "non-native")
+  if (nrow(db_inv) != 0) {
     expect_equal(species_nativity(species2, database_inventory = db_inv), "native")
     expect_message(species_nativity("fake_species", database_inventory = db_inv),
                    "Species not found in specified database.")
-  } else {
-    # for when database download fails
-    expect_message(species_nativity(species, 149))
   }
 })
