### 2024-04-11 15956d8
improved testing when API fails
