+++ tests/testthat/test-transect_list_glance.R
@@ -17,9 +17,7 @@ test_that("transect_list_glance works", {
     test_df <- transect_list_glance(test_list)
 
     expect_equal(ncol(test_df), 54)
-    expect_equal(nrow(test_df), length(test_list))
     expect_equal(typeof(test_df$total_mean_c), "double")
-    expect_equal(nrow(test_df), 2)
   }
 })
 
