### 2024-04-11 8dc9c56
update news for 0.5.3 release
