+++ NEWS.md
@@ -1,4 +1,4 @@
-# fqar 0.5.2
+# fqar 0.5.2 and 0.5.3
 
 * increased stability when universalFQA fails, which is becoming more common over time.
 
