### 2024-04-13 38417f6
Clarify role of indexing functions
