### 2024-04-13 665ec03
only import txtProgressBar from utils
