+++ NAMESPACE
@@ -34,7 +34,6 @@ import(dplyr)
 import(httr)
 import(jsonlite)
 import(tidyr)
-import(utils)
 importFrom(ggplot2,aes)
 importFrom(ggplot2,geom_col)
 importFrom(ggplot2,geom_vline)
@@ -49,3 +48,4 @@ importFrom(memoise,memoise)
 importFrom(rlang,.data)
 importFrom(rlang,is_interactive)
 importFrom(stats,sd)
+importFrom(utils,txtProgressBar)
