+++ man/assessment_glance.Rd
@@ -14,58 +14,58 @@ assessment_glance(data_set)
 \value{
 A data frame with 52 columns:
 \itemize{
-   title (character)
-   date (date)
-   site_name (character)
-   city (character)
-   county (character)
-   state (character)
-   country (character)
-   fqa_db_region (character)
-   fqa_db_publication_year (character)
-   fqa_db_description (character)
-   custom_fqa_db_name (character)
-   custom_fqa_db_description (character)
-   practitioner (character)
-   latitude (character)
-   longitude (character)
-   weather_notes (character)
-   duration_notes (character)
-   community_type_notes (character)
-   other_notes (character)
-   private_public (character)
-   total_mean_c (numeric)
-   native_mean_c (numeric)
-   total_fqi (numeric)
-   native_fqi (numeric)
-   adjusted_fqi (numeric)
-   c_value_zero (numeric) Percent of c-values 0
-   c_value_low (numeric) Percent of c-values 1-3
-   c_value_mid (numeric) Percent of c-values 4-6
-   c_value_high (numeric) Percent of c-values 7-10
-   native_tree_mean_c (numeric)
-   native_shrub_mean_c (numeric)
-   native_herbaceous_mean_c (numeric)
-   total_species (numeric)
-   native_species (numeric)
-   non_native_species (numeric)
-   mean_wetness (numeric)
-   native_mean_wetness (numeric)
-   tree (numeric)
-   shrub (numeric)
-   vine (numeric)
-   forb (numeric)
-   grass (numeric)
-   sedge (numeric)
-   rush (numeric)
-   fern (numeric)
-   bryophyte (numeric)
-   annual (numeric)
-   perennial (numeric)
-   biennial (numeric)
-   native_annual (numeric)
-   native_perennial (numeric)
-   native_biennial (numeric)
+   \item title (character)
+   \item date (date)
+   \item site_name (character)
+   \item city (character)
+   \item county (character)
+   \item state (character)
+   \item country (character)
+   \item fqa_db_region (character)
+   \item fqa_db_publication_year (character)
+   \item fqa_db_description (character)
+   \item custom_fqa_db_name (character)
+   \item custom_fqa_db_description (character)
+   \item practitioner (character)
+   \item latitude (character)
+   \item longitude (character)
+   \item weather_notes (character)
+   \item duration_notes (character)
+   \item community_type_notes (character)
+   \item other_notes (character)
+   \item private_public (character)
+   \item total_mean_c (numeric)
+   \item native_mean_c (numeric)
+   \item total_fqi (numeric)
+   \item native_fqi (numeric)
+   \item adjusted_fqi (numeric)
+   \item c_value_zero (numeric) Percent of c-values 0
+   \item c_value_low (numeric) Percent of c-values 1-3
+   \item c_value_mid (numeric) Percent of c-values 4-6
+   \item c_value_high (numeric) Percent of c-values 7-10
+   \item native_tree_mean_c (numeric)
+   \item native_shrub_mean_c (numeric)
+   \item native_herbaceous_mean_c (numeric)
+   \item total_species (numeric)
+   \item native_species (numeric)
+   \item non_native_species (numeric)
+   \item mean_wetness (numeric)
+   \item native_mean_wetness (numeric)
+   \item tree (numeric)
+   \item shrub (numeric)
+   \item vine (numeric)
+   \item forb (numeric)
+   \item grass (numeric)
+   \item sedge (numeric)
+   \item rush (numeric)
+   \item fern (numeric)
+   \item bryophyte (numeric)
+   \item annual (numeric)
+   \item perennial (numeric)
+   \item biennial (numeric)
+   \item native_annual (numeric)
+   \item native_perennial (numeric)
+   \item native_biennial (numeric)
 }
 }
 \description{
