### 2024-04-13 equitable-equations/fqar#8 @equitable-equations
@ifoxfoot thanks for these helpful notes. I've integrated all your suggestions as outlined below.

** README

- It looks like there are two periods after the link to the universal calculator in the first paragraph. **FIxed.**

- You present four types of functions: indexing, downloading, tidying, and analytic. For the last three types there is some text explaining what the functions do interspersed with code blocks, but for the indexing functions there is only code. **Explanatory text has been added at the beginning of this section.**

- Community guidelines. **Added How to Contribute to README**.

** Vignette

- There are only two authors listed on the vignette but there are three on the paper/R package. Just wondering if this is on purpose. **Yes, this is intentional. The third author joined the project later, after the vignette had been written.**

** R package documentation

- In the description file lines 11-13, you could change the description to specify that you are downloading floristic quality assessment data from the universal FQA calculator. **The Freyman reference is included for this purpose. I'd prefer to leave the universal FQA link out of the description because it has occasionally failed in the past, throwing an annoying CRAN note. I hope this is ok! The link is found in most documentation pages.**

- Documenting what data goes in each column of your function outputs would be helpful for all of your functions. Some column names are self explanatory but some are not. For example, what is discrepancy_c in the output of assessment_cooccurrences_summary? **I agree and have added this to the to-do list for version 0.6.0, planned for this summer. This is a larger task since these data frames include so many columns and there are many such functions. I hope it's possible for JOSS to move forward without waiting for this change to be implemented.**

- In the function documentation of assessment_glance the description of the data returned is different from all of your other functions. In your other function documentation you write. **This oversight has been fixed.**

** General documentation comments

- One thing you could mention is who hosts the Universal FQA Calculator database. **Added to README**

- For arguments like database_id, you mention that the ID is specified by the universal calculator. Is there any risk that these IDs would change over time? **These IDs are assigned by the system when the entries are first added by the user, so they can reasonably be expected to be static. I haven't encountered any such changes and don't anticipate any.**
