### 2024-04-13 equitable-equations/fqar#10 @equitable-equations
I've manually added a section to the readme and will learn about `usethis::use_tidy_contributing()` for the next update, scheduled for the summer. Cool stuff!
