### 2024-04-13 equitable-equations/fqar#9 @equitable-equations
Done. I also had this issue while developing the package but incorrectly assumed it as specific to `devtools`. I'll push all changes to the repo shortly.
