### 2024-04-13 equitable-equations/fqar#7 @equitable-equations
I've re-tested everything carefully and things appear to be ok. I'm closing this issue.
