### 2024-04-13 equitable-equations/fqar#6 @equitable-equations
@ifoxfoot yes to all. Specifically:

- I've added the word "generally" to allow for the unusual handling of non-native species

- The paragraph you reference has been recast, ending "...though it has so far gone largely unexplored due to a lack of both technical tools for interacting programmatically with the repository and accessible workflows for analyzing the floristic quality data housed there."

- I'm hoping to tackle iNaturalist next!

All changes will be pushed to the repo shortly.
