+++ paper/paper.md
@@ -26,9 +26,9 @@ bibliography: fqar_refs.bib
 
 # Summary
 
-Floristic Quality Assessment (FQA) is a standardized method for rating the ecological value of natural areas based on the plant species found within them [@spyreas2019floristic; @swink1994plants]. Each species considered native to a particular region is assigned a *coefficient of conservatism*, C, on a scale of 0-10 by experts in local flora (non-native species are assigned a zero C-value by default). Larger values of C correspond to species that tend to be found in undegraded sites, while lower values indicate species that are more tolerant to human impacts [@bauer]. An inventory of the site is conducted and the average of the C-values found there is computed. This *native mean C-value*, sometimes weighted by the total number of plant species identified to give the so-called *floristic quality index* [@bowles2006testing], is frequently used by land managers and other agents to quantify an area's state of conservancy [@zinnen]. 
+Floristic Quality Assessment (FQA) is a standardized method for rating the ecological value of natural areas based on the plant species found within them [@spyreas2019floristic; @swink1994plants]. Each species known to be found in a particular region is assigned a *coefficient of conservatism*, C, on a scale of 0-10 by experts in local flora (non-native species are generally assigned a zero C-value by default). Larger values of C correspond to species that tend to be found in undegraded sites, while lower values indicate species that are more tolerant to human impacts [@bauer]. An inventory of the site is conducted and the average of the C-values found there is computed. This *native mean C-value*, sometimes weighted by the total number of plant species identified to give the so-called *floristic quality index* [@bowles2006testing], is frequently used by land managers and other agents to quantify an area's state of conservancy [@zinnen]. 
 
-In recent years, it has become increasingly standard for practitioners to upload their floristic quality assessments to a central repository, [universalfqa.org](https://universalfqa.org/) [@freyman2016universal], which already includes tens of thousands of assessments from over one hundred floristic quality databases. This large public data cache represents a potentially invaluable resource for quantitative ecologists, though it has so far gone largely unexplored due to a lack of both technical tools for interacting programatically with the repository and statistical methods for analyzing the floristic quality data housed there. 
+In recent years, it has become increasingly standard for practitioners to upload their floristic quality assessments to a central repository, [universalfqa.org](https://universalfqa.org/) [@freyman2016universal], which already includes tens of thousands of assessments from over one hundred floristic quality databases. This large public data cache represents a potentially invaluable resource for quantitative ecologists, though it has so far gone largely unexplored due to a lack of both technical tools for interacting programmatically with the repository and accessible workflows for analyzing the floristic quality data housed there. 
 
 `fqar` is an R [@rcore] package which facilitates the analysis of occurrence and co-occurrence of plant taxa at the regional level. Pulling data on-demand from  [universalfqa.org](https://universalfqa.org/), it provides both organizational tools for handling the disparate sorts of data housed there and statistical ones for drawing novel conclusions from that data. 
 
@@ -36,11 +36,11 @@ In recent years, it has become increasingly standard for practitioners to upload
 
 The [universalfqa.org](https://universalfqa.org/) website is calibrated for practitioners in the field rather than data analysts at their desks. It facilitates the recording, storing, and publicizing of individual floristic quality assessments and performs calculations of the statistical measures most often cited by land managers and conservation organizations in their reporting, including native mean-C. However, its focus on individual assessments is not well-suited to analyses that might wish to consider multiple assessments simultaneously.
 
-The `fqar` package enables analysis with a wider lens, allowing users to consider database-wide records of plant taxa or characteristics. By examining entire collections of assessments simultaneously, ecologists may gain insights into floristic quality assessment as well as the various plant species it tracks. Among the wide range of questions made answerable by `fqar` are the following:
[…]
