### 2024-04-13 #6366 @equitable-equations
@ifoxfoot @mhesselbarth thanks again for your patience. This package has been updated to account for the unexpected API failures seen last month, changes which have been pushed to the main branch of the repo and which are now on CRAN. I believe i have responded to all of your suggestions and comments, which were both constructive and on-point. A new version of the paper reflecting all of this can be found in the `joss revision` branch.
