+++ cran-comments.md
@@ -1,14 +1,8 @@
 ## R CMD check results
 
-0 errors | 0 warnings | 0 notes*
+0 errors | 0 warnings | 0 notes
 
-* rhub::check_for_cran sometimes produces the following notes, which I understand to be noise:
-
-1. Found the following files/directories: 'lastMiKTeXException'
-
-2. Checking for non-standard things in the check directory ''NULL''
-
-* A website referred to in the documentation has been down occasionally in recent weeks. This may possibly produce another note. I've chose to leave the link in the help files for added transparency. 
+My apologies for the pushing an additional update so quickly. An error appeared unexpectedly at https://cran.r-project.org/web/checks/check_results_fqar.html which I've now corrected.
 
 
 There are currently no downstream dependencies for this package.
