+++ DESCRIPTION
@@ -1,7 +1,7 @@
 Package: fqar
 Type: Package
 Title: Floristic Quality Assessment Tools for R
-Version: 0.5.2
+Version: 0.5.3
 Authors@R: c(
   person("Andrew", "Gard", , email = "agard@lakeforest.edu", role = c("aut", "cre"),
     comment = c(ORCID = "0000-0003-4434-0755")),
