+++ NAMESPACE
@@ -48,4 +48,5 @@ importFrom(memoise,memoise)
 importFrom(rlang,.data)
 importFrom(rlang,is_interactive)
 importFrom(stats,sd)
+importFrom(utils,tail)
 importFrom(utils,txtProgressBar)
