+++ cran-comments.md
@@ -1,8 +1,9 @@
 ## R CMD check results
 
-0 errors | 0 warnings | 0 notes
+0 errors | 0 warnings | 1 note*
 
-My apologies for the pushing an additional update so quickly. An error appeared unexpectedly at https://cran.r-project.org/web/checks/check_results_fqar.html which I've now corrected.
+* Days since last update: 4
+My apologies for the pushing an additional update so quickly. An error appeared unexpectedly at https://cran.r-project.org/web/checks/check_results_fqar.html. This was due to a testing oversight which I've now corrected. 
 
 
 There are currently no downstream dependencies for this package.
