+++ R/download_assessment_list.R
@@ -47,6 +47,7 @@ download_assessment_list <- function(database_id, ...) {
 
   if (nrow(inventories_summary) == 0){
     return(invisible(list()))
+    message("No data returned for specified database.")
   }
 
   inventories_requested <- inventories_summary |>
@@ -68,6 +69,7 @@ download_assessment_list <- function(database_id, ...) {
       width = length(inventories_requested$id),
       char = "="
     )
+
     for (i in seq_along(inventories_requested$id)) {
       results[[i]] <-  download_assessment(inventories_requested$id[i])
       utils::setTxtProgressBar(pb, i)
@@ -80,6 +82,7 @@ download_assessment_list <- function(database_id, ...) {
 
   if (length(results) == 0){
     message("No matches found. Empty list returned.")
+    return(invisible(results))
   }
 
   results
