+++ cran-comments.md
@@ -3,7 +3,7 @@
 0 errors | 0 warnings | 1 note*
 
 * Days since last update: 4
-My apologies for the pushing an additional update so quickly. An error appeared unexpectedly at https://cran.r-project.org/web/checks/check_results_fqar.html. This was due to a testing oversight which I've now corrected. 
+My apologies for the pushing an additional update so quickly. An error appeared unexpectedly at https://cran.r-project.org/web/checks/check_results_fqar.html. This was due to a minor testing oversight which I've now corrected. 
 
 
 There are currently no downstream dependencies for this package.
