### 2024-04-14 c33bd2f
add fqacalc reference
