+++ paper/paper.md
@@ -36,7 +36,7 @@ In recent years, it has become increasingly standard for practitioners to upload
 
 The [universalfqa.org](https://universalfqa.org/) website is calibrated for practitioners in the field rather than data analysts at their desks. It facilitates the recording, storing, and publicizing of individual floristic quality assessments and performs calculations of the statistical measures most often cited by land managers and conservation organizations in their reporting, including native mean-C. However, its focus on individual assessments is not well-suited to analyses that might wish to consider multiple assessments simultaneously.
 
-This package compliments existing R packages for floristic quality analysis, including `fqacalc` and `fqadata`, which support the work of field practitioners wishing to make use of R. The `fqar` package enables analysis with a wider lens, allowing users to consider database-wide records of plant taxa or characteristics. By examining entire collections of assessments simultaneously, ecologists may gain insights into floristic quality assessment as well as the various plant species it tracks. Among the wide range of questions made answerable by `fqar` are the following:
+This package compliments existing R packages for floristic quality analysis, including `fqacalc` and `fqadata`, which support the work of field practitioners wishing to make use of R [@fqacalc; @fqadata]. The `fqar` package enables analysis with a wider lens, allowing users to consider database-wide records of plant taxa or characteristics. By examining entire collections of assessments simultaneously, ecologists may gain insights into floristic quality assessment as well as the various plant species it tracks. Among the wide range of questions made answerable by `fqar` are the following:
 
 - what is the co-occurrence profile of a given species of interest? What other plants (or types of plants) is it most frequently identified alongside?
 
