+++ paper/fqar_refs.bib
@@ -1,12 +1,29 @@
 %% This BibTeX bibliography file was created using BibDesk.
 %% https://bibdesk.sourceforge.io/
 
-%% Created for Andrew Gard at 2023-08-12 13:59:04 -0500 
+%% Created for Andrew Gard at 2024-04-14 10:37:20 -0500 
 
 
 %% Saved with string encoding Unicode (UTF-8) 
 
 
+
+@manual{fqadata,
+	author = {Iris Foxfoot},
+	date-added = {2024-04-14 10:36:45 -0500},
+	date-modified = {2024-04-14 10:37:16 -0500},
+	title = {fqadata: Contains Regional Floristic Quality Assessment Databases},
+	url = {https://CRAN.R-project.org/package=fqadata},
+	year = {2023}}
+
+@manual{fqacalc,
+	author = {Iris Foxfoot},
+	date-added = {2024-04-14 10:34:05 -0500},
+	date-modified = {2024-04-14 10:35:36 -0500},
+	title = {fqacalc: Calculate Floristic Quality Assessment Metrics},
+	url = {https://CRAN.R-project.org/package=fqacalc},
+	year = {2023},
+	bdsk-url-1 = {https://CRAN.R-project.org/package=fqacalc}}
 
 @article{JSSv059i10,
 	abstract = {A huge amount of effort is spent cleaning data to get it ready for analysis, but there has been little research on how to make data cleaning as easy and effective as possible. This paper tackles a small, but important, component of data cleaning: data tidying. Tidy datasets are easy to manipulate, model and visualize, and have a specific structure: each variable is a column, each observation is a row, and each type of observational unit is a table. This framework makes it easy to tidy messy datasets because only a small set of tools are needed to deal with a wide range of un-tidy datasets. This structure also makes it easier to develop tidy tools for data analysis, tools that both input and output tidy datasets. The advantages of a consistent data structure and matching tools are demonstrated with a case study free from mundane data manipulation chores.},
