### 2024-04-17 #6366 @mikemahoney218
@ifoxfoot , would you be able to look at the changes @equitable-equations has made, and let me know if the current state of the software addresses your initial concerns (or open new issues on the repository if you still have items you can't check on the checklist)?
