### 2024-04-18 #6366 @mikemahoney218
Thank you so much @ifoxfoot , and thank you for reviewing for JOSS!

 @equitable-equations , I'll do my final checks and start the process for recommending acceptance tomorrow.
