### 2024-04-18 e0126a8
remove paper workflow from main branch
