+++ paper/fqar_refs.bib
@@ -1,7 +1,7 @@
 %% This BibTeX bibliography file was created using BibDesk.
 %% https://bibdesk.sourceforge.io/
 
-%% Created for Andrew Gard at 2024-04-14 10:37:20 -0500 
+%% Created for Andrew Gard at 2024-04-20 10:22:56 -0500 
 
 
 %% Saved with string encoding Unicode (UTF-8) 
@@ -11,16 +11,17 @@
 @manual{fqadata,
 	author = {Iris Foxfoot},
 	date-added = {2024-04-14 10:36:45 -0500},
-	date-modified = {2024-04-14 10:37:16 -0500},
-	title = {fqadata: Contains Regional Floristic Quality Assessment Databases},
+	date-modified = {2024-04-20 08:45:54 -0500},
+	title = {{fqadata}: Contains Regional Floristic Quality Assessment Databases},
 	url = {https://CRAN.R-project.org/package=fqadata},
-	year = {2023}}
+	year = {2023},
+	bdsk-url-1 = {https://CRAN.R-project.org/package=fqadata}}
 
 @manual{fqacalc,
 	author = {Iris Foxfoot},
 	date-added = {2024-04-14 10:34:05 -0500},
-	date-modified = {2024-04-14 10:35:36 -0500},
-	title = {fqacalc: Calculate Floristic Quality Assessment Metrics},
+	date-modified = {2024-04-20 08:45:42 -0500},
+	title = {{fqacalc}: Calculate Floristic Quality Assessment Metrics},
 	url = {https://CRAN.R-project.org/package=fqacalc},
 	year = {2023},
 	bdsk-url-1 = {https://CRAN.R-project.org/package=fqacalc}}
@@ -44,23 +45,25 @@
 @article{matthews2015null,
 	author = {Matthews, Jeffrey W and Spyreas, Greg and Long, Colleen M},
 	date-added = {2023-08-11 13:27:00 -0500},
-	date-modified = {2023-08-11 13:27:00 -0500},
+	date-modified = {2024-04-20 08:56:53 -0500},
 	journal = {Ecological Indicators},
 	pages = {1--7},
 	publisher = {Elsevier},
 	title = {A null model test of Floristic Quality Assessment: Are plant species' Coefficients of Conservatism valid?},
+	url = {https://doi.org/10.1016/j.ecolind.2014.11.017},
 	volume = {52},
 	year = {2015}}
 
 @article{bourdaghs2006properties,
 	author = {Bourdaghs, Michael and Johnston, Carol A and Regal, Ronald R},
 	date-added = {2023-08-11 12:48:37 -0500},
-	date-modified = {2023-08-11 12:48:37 -0500},
+	date-modified = {2024-04-20 08:54:24 -0500},
 	journal = {Wetlands},
 	number = {3},
 	pages = {718--735},
 	publisher = {Springer},
 	title = {Properties and performance of the floristic quality index in Great Lakes coastal wetlands},
+	url = {https://doi.org/10.1672/0277-5212(2006)26[718:PAPOTF]2.0.CO;2},
 	volume = {26},
 	year = {2006}}
 
@@ -71,23 +74,6 @@
 	organization = {Indiana Academy of Science Indianapolis},
 	title = {Flora of the Chicago region: A floristic and ecological synthesis},
 	year = {2017}}
-
-@article{oldham1995floristic,
-	author = {Oldham, MJ and Bakowsky, WD and Sutherland, DA},
-	date-added = {2022-08-30 10:59:12 -0500},
-	date-modified = {2022-08-30 10:59:12 -0500},
-	journal = {Ontario. Natural Heritage Information Center Report, Ministry of Natural Resources, Peterborough, Ontario, Canada},
[…]
