### 2024-04-20 7381881
add dois and update refs
