### 2024-04-20 #6366 @equitable-equations
@mikemahoney218 I think I've done all of this. Hopefully the .bib file clears your checks now!

- [x] Go through the missing DOIs listed above and add the ones which are correct. Note that our systems will need all the references in your .bib to be correct before we can accept the paper; if you get rid of items in the .bib that aren't used (for instance, I don't think you cite dplyr) it might make things easier. **Done**

- [x] Add curly braces {} around fqacalc and fqadata in your .bib, so that their titles are formatted properly in the references list. **Done**

Archive the reviewed software in Zenodo or a similar service (e.g., figshare, an institutional repository). Make sure that the license listed for the archive is the same as the software license! **Done** The Zenodo archive can be seen [here](https://doi.org/10.5281/zenodo) and the GitHub release [here](https://github.com/equitable-equations/fqar/releases/tag/v0.5.3). Both are tagged as v0.5.3.

- [x] Check the archival deposit (e.g., in Zenodo) has the correct metadata. This includes the title (should match the paper title) and author list (make sure the list is correct and people who only made a small fix are not on it). You may also add the authors' ORCID. **Done**

- [x] Please list the DOI of the archived version here. The DOI is **https://doi.org/10.5281/zenodo.11002086**
