### 2024-04-22 #6366 @mikemahoney218
@equitable-equations , would you be willing to update the Zenodo archive to list the paper co-authors as co-authors on the release, rather than "project members"? Or, do you have a reason that they shouldn't be listed as authors on this archive? I think this is the last thing outstanding before I can go ahead and recommend acceptance, and hand this back to the EiC for final processing.
